#32: Created new directory and moved files
